requirements.txt
requests==2.32.3
Flask==3.0.3
pandas==2.2.3
scikit-learn==1.5.2
python-dotenv==1.0.1
gunicorn==23.0.0
Flask-APScheduler==1.13.1
python-dateutil~=2.9.0.post0